+++ dotnet_exercise_csharp_5_garage_1/Classes/Car.cs
@@ -4,13 +4,13 @@ namespace dotnet_exercise_csharp_5_garage_1.Classes
 {
     internal class Car : Vehicle
     {
-        private double _cylinderVolume;
+        private uint _cylinderVolume;
         private string _fuleType;
 
-        public double CylinderVolume => _cylinderVolume;
+        public uint CylinderVolume => _cylinderVolume;
         public string FuleType => _fuleType;
 
-        public Car(string regNbr, string color, uint wheels, string fuleType, double cylinderVolume)
+        public Car(string regNbr, string color, uint wheels, string fuleType, uint cylinderVolume)
             : base(regNbr, color, wheels)
         {
             _fuleType = fuleType;
