+++ dotnet_exercise_csharp_5_garage_1/Classes/Motorcycle.cs
@@ -6,13 +6,13 @@ namespace dotnet_exercise_csharp_5_garage_1.Classes
 {
     internal class Motorcycle : Vehicle
     {
-        private double _cylinderVolume;
+        private uint _cylinderVolume;
         private uint _seats;
 
-        public double CylinderVolume => _cylinderVolume;
+        public uint CylinderVolume => _cylinderVolume;
         public uint Seats => _seats;
 
-        public Motorcycle(string regNbr, string color, uint wheels, uint nbrOfSeats, double cylinderVolume)
+        public Motorcycle(string regNbr, string color, uint wheels, uint nbrOfSeats, uint cylinderVolume)
             : base(regNbr, color, wheels)
         {
             _cylinderVolume = cylinderVolume;
